[#7] Rename: pages 폴더 추가
